+++ docs/rules/Proj1010.md
@@ -1,0 +1,16 @@
+---
+parent: Analyzers
+ancestor: Rules
+permalink: /rules/Proj1010
+---
+
+# Proj1010: Avoid enabling deprecated rules
+Diagnostic rules that are being depreacted will eventually be dropped. Therefore
+it is advised not enable them if they are not configured to be enabled by default.
+
+## Non-compliant
+``` ini
+is_global = true
+
+dotnet_diagnostic.S1145.severity = warning
+```
