+++ src/DotNetProjectFile.Analyzers/Analyzers/Helpers/RoslynRules.cs
@@ -12,7 +12,7 @@ public static class RoslynRules
         .Split(';')
         .ToFrozenSet();
 
-    public static readonly FrozenSet<string> Obsolete = "AV0000;CA1714;CA1717;CA1801;CA2109;MA0034;MEN020;Proj1000;Proj9999;RCS0004;RCS0011i;RCS0014;RCS0015i;RCS0022;RCS0027i;RCS0028i;RCS0032i;RCS0038;RCS0043;RCS0047;RCS0051i;RCS0052i;RCS1008;RCS1009;RCS1010;RCS1012;RCS1014a;RCS1014i;RCS1016a;RCS1016b;RCS1016i;RCS1017;RCS1018i;RCS1022;RCS1022FadeOut;RCS1024;RCS1025;RCS1026;RCS1027;RCS1028;RCS1029;RCS1030;RCS1035;RCS1036;RCS1036a;RCS1038;RCS1040;RCS1041;RCS1045;RCS1045a;RCS1050i;RCS1051a;RCS1064;RCS1065;RCS1066;RCS1067;RCS1072;RCS1078i;RCS1086;RCS1087;RCS1088;RCS1090i;RCS1091;RCS1092;RCS1096i;RCS1100;RCS1101;RCS1104a;RCS1106;RCS1122;RCS1153;RCS1176;RCS1177;RCS1184;RCS1185;RCS1207i;RCS1219;RCS1246a;RCS1248a;RCS1248i;ROS0002;RS1007;RS1015;RS1020;RS1028;S1145;S1227;S1697;S2070;S2228;S2278;S2353;S2373;S2387;S2758;S3693;S3884;S3900;S4142;S4212;S4214;S4432;S4564;S5042;S6803;SA0000"
+    public static readonly FrozenSet<string> Deprecated = "AV0000;CA1714;CA1717;CA1801;CA2109;MA0034;MEN020;Proj1000;Proj9999;RCS0004;RCS0011i;RCS0014;RCS0015i;RCS0022;RCS0027i;RCS0028i;RCS0032i;RCS0038;RCS0043;RCS0047;RCS0051i;RCS0052i;RCS1008;RCS1009;RCS1010;RCS1012;RCS1014a;RCS1014i;RCS1016a;RCS1016b;RCS1016i;RCS1017;RCS1018i;RCS1022;RCS1022FadeOut;RCS1024;RCS1025;RCS1026;RCS1027;RCS1028;RCS1029;RCS1030;RCS1035;RCS1036;RCS1036a;RCS1038;RCS1040;RCS1041;RCS1045;RCS1045a;RCS1050i;RCS1051a;RCS1064;RCS1065;RCS1066;RCS1067;RCS1072;RCS1078i;RCS1086;RCS1087;RCS1088;RCS1090i;RCS1091;RCS1092;RCS1096i;RCS1100;RCS1101;RCS1104a;RCS1106;RCS1122;RCS1153;RCS1176;RCS1177;RCS1184;RCS1185;RCS1207i;RCS1219;RCS1246a;RCS1248a;RCS1248i;ROS0002;RS1007;RS1015;RS1020;RS1028;S1145;S1227;S1697;S2070;S2228;S2278;S2353;S2373;S2387;S2758;S3693;S3884;S3900;S4142;S4212;S4214;S4432;S4564;S5042;S6803;SA0000"
         .Split(';')
         .ToFrozenSet();
 }
