+++ src/DotNetProjectFile.Analyzers/Rule.cs
@@ -1366,15 +1366,15 @@ public static partial class Rule
         tags: ["suppression", "severity", "warning", "error"],
         category: Category.Configuration);
 
-    public static DiagnosticDescriptor AvoidEnablingObsoleteRules => New(
-          id: 1010,
-          title: "Avoid enabling obsolete rules",
-          message: "Rule {0} is obsolete and should not be enabled",
-          description:
-              "Enabling obsolete rules has no effect, as they are no longer " +
-              "evaluated by the compiler.",
-          tags: [".globalconfig", "diagnostic", "obsolete"],
-          category: Category.Configuration);
+    public static DiagnosticDescriptor AvoidEnablingDeprecatedRules => New(
+        id: 1010,
+        title: "Avoid enabling deprecated rules",
+        message: "Rule {0} is deprecated and should not be enabled",
+        description:
+            "Enabling deprecated rules has no effect, as they are no longer " +
+            "evaluated by the compiler.",
+        tags: [".globalconfig", "diagnostic", "deprecated"],
+        category: Category.Configuration);
 
 
     public static DiagnosticDescriptor AvoidUsingMoq => New(
